added new stylesheet file and relevant changes to SettingsScreen, also changed Button component into Pressable for more styling options.
